% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_bands.R
\name{gt_bands}
\alias{gt_bands}
\title{Download all available bands}
\usage{
gt_bands(product)
}
\arguments{
\item{product}{a valid GeoNEX product name}
}
\value{
A data frame of all available bands for a GeoNEX
Products Subsets products
}
\description{
Lists all available bands for a GeoNEX Products Subset product.
}
\examples{

\donttest{
# list all available GeoNEX  Products Subsets products
bands <- gt_bands(product = "geonex_GO16_L1G")
head(bands)

}

}
\seealso{
\code{\link[GeoNEXTools]{gt_products}}
\code{\link[GeoNEXTools]{gt_sites}}
}
