% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_batch_subset.R
\name{gt_batch_subset}
\alias{gt_batch_subset}
\title{Batch download GeoNEX Products subsets}
\usage{
gt_batch_subset(
  df,
  product,
  band,
  start = "2016-01-01",
  end = format(Sys.time(), "\%Y-\%m-\%d"),
  km_lr = 0,
  km_ab = 0,
  out_dir = tempdir(),
  internal = TRUE
)
}
\arguments{
\item{df}{a CSV file or data frame holding site IDs
When providing a CSV make sure that the
data are comma separated.}

\item{product}{a valid GeoNEX product name}

\item{band}{band to download}

\item{start}{start date}

\item{end}{end date}

\item{km_lr}{km left-right to sample}

\item{km_ab}{km above-below to sample}

\item{out_dir}{location where to store all data}

\item{internal}{should the data be returned as an internal data structure
\code{TRUE} or \code{FALSE} (default = \code{TRUE})}
}
\value{
A data frame combining meta-data and actual data values, data from
different sites is concatenated into one large dataframe. Subsets can be
created by searching on sitename.
}
\description{
Lists all available dates for a GeoNEX Products Subset product
at a particular location.
}
\examples{

\dontrun{
# create data frame with a site_id column
df <- data.frame("site_id" = c("us_idaho_twrs",
                              "us_illinois_bondville"))

print(df)

# test batch download
subsets <- gt_batch_subset(df = df,
                        product = "geonex_GO16_L1G",
                        band = "BAND01",
                        internal = TRUE,
                        start = "2004-01-01",
                        end = "2004-03-31")

# the same can be done using a CSV file with
# a data structure similar to the dataframe above

write.table(df, file.path("my_sites.csv"),
 quote = FALSE,
 row.names = FALSE,
 col.names = TRUE,
 sep = ",")

# test batch download form CSV
subsets <- gt_batch_subset(df = file.path("my_sites.csv"),
                        product = "geonex_GO16_L1G",
                        band = "BAND01",
                        internal = TRUE,
                        start = "2004-01-01",
                        end = "2004-03-31"
                        )

head(subsets)
}
}
\seealso{
\code{\link[GeoNEXTools]{gt_sites}}
\code{\link[GeoNEXTools]{gt_bands}}
\code{\link[GeoNEXTools]{gt_products}}
\code{\link[GeoNEXTools]{gt_subset}}
}
