% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_products.R
\name{gt_products}
\alias{gt_products}
\title{Download all available products}
\usage{
gt_products()
}
\value{
A data frame of all available GeoNEX Products Subsets products
}
\description{
Lists all available GeoNEX Products Subset products.
}
\examples{

\donttest{
# list all available GeoNEX Products Subsets products
products <- gt_products()
head(products)
}

}
\seealso{
\code{\link[GeoNEXTools]{gt_bands}}
\code{\link[GeoNEXTools]{gt_sites}}
}
