% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_sites.R
\name{gt_sites}
\alias{gt_sites}
\title{Download all available fixed sites}
\usage{
gt_sites(network)
}
\arguments{
\item{network}{the network for which to generate the site list,
when not provided the complete list is provided}
}
\value{
A data frame of all available GeoNEX Products Subsets
pre-processed sites
}
\description{
Lists all available GeoNEX Products Subset pre-processed sites
}
\examples{

\donttest{
# list all available GeoNEX Products Subsets products
sites <- gt_sites()
print(head(sites))
}

}
\seealso{
\code{\link[GeoNEXTools]{gt_products}}
\code{\link[GeoNEXTools]{gt_bands}}
}
