% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_subset.R
\name{gt_subset}
\alias{gt_subset}
\title{Download GeoNEX Products subsets}
\usage{
gt_subset(
  product,
  band,
  start = "2016-01-01",
  end = format(Sys.time(), "\%Y-\%m-\%d"),
  km_lr = 0,
  km_ab = 0,
  site_id,
  network,
  site_name = "sitename",
  out_dir = tempdir(),
  internal = TRUE
)
}
\arguments{
\item{product}{a valid GeoNEX product name}

\item{band}{band or bands (as a character vector) to download}

\item{start}{start date}

\item{end}{end date}

\item{km_lr}{km left-right to sample (rounded to the nearest integer)}

\item{km_ab}{km above-below to sample (rounded to the nearest integer)}

\item{site_id}{site id}

\item{network}{the network for which to generate the site list,
when not provided the complete list is provided}

\item{site_name}{arbitrary site name used in writing data to file
(default = sitename)}

\item{out_dir}{path where to store the data if writing to disk
(default = tempdir())}

\item{internal}{should the data be returned as an internal data structure
\code{TRUE} or \code{FALSE} (default = \code{TRUE})}
}
\value{
A data frame combining meta-data and actual data values.
}
\description{
Download a GeoNEX Products Subset product
for a given GeoNEX subset sites.
}
\examples{

\donttest{
# list all available GeoNEX Products Subsets products
# download data
subset <- gt_subset(product = "geonex_GO16_L1G",
                        band = "BAND01",
                        start = "2020-01-01",
                        end = "2020-03-31",
                        site_id = "us_alabama_al_fairhope_3_ne")
 head(subset)
}
}
\seealso{
\code{\link[GeoNEXTools]{gt_sites}}
\code{\link[GeoNEXTools]{gt_bands}}
\code{\link[GeoNEXTools]{gt_products}}
\code{\link[GeoNEXTools]{gt_batch_subset}}
}
